New exploit for GlusterFS CVE-2018-1088 and CVE-2018-1112
